+++ src/apis/member/queries.ts
@@ -9,16 +9,4 @@ export const member = {
     queryKey: ['check', accessToken],
     queryFn: () => memberApi.check(accessToken),
   }),
-  info: (uuid: string) => ({
-    queryKey: ['info', uuid],
-    queryFn: () => memberApi.info(uuid),
-  }),
-  times: (uuid: string) => ({
-    queryKey: ['times', uuid],
-    queryFn: () => memberApi.times(uuid),
-  }),
-  participants: (uuid: string) => ({
-    queryKey: ['participants', uuid],
-    queryFn: () => memberApi.participants(uuid),
-  }),
 } as const;
